moved SlotManager into it's own file, ValueSlot back into base_slots, changed some imports
